feat: 04-authentication

+++ controllers/contactsControllers.js
@@ -2,13 +2,15 @@ import * as contactsService from "../services/contactsServices.js";
 import HttpError from "../helpers/HttpError.js";
 
 export const getAllContacts = async (req, res) => {
-  const contacts = await contactsService.listContacts();
+  const { id: owner } = req.user;
+  const contacts = await contactsService.listContacts(owner);
   res.json(contacts);
 };
 
 export const getOneContact = async (req, res) => {
   const { id } = req.params;
-  const contact = await contactsService.getContactById(id);
+  const { id: owner } = req.user;
+  const contact = await contactsService.getContact({ id, owner });
   if (!contact) {
     throw HttpError(404, `Contact with id=${id} not found`);
   }
@@ -18,7 +20,8 @@ export const getOneContact = async (req, res) => {
 
 export const deleteContact = async (req, res) => {
   const { id } = req.params;
-  const deletedContact = await contactsService.removeContact(id);
+  const { id: owner } = req.user;
+  const deletedContact = await contactsService.removeContact({ id, owner });
   if (!deletedContact) {
     throw HttpError(404, `Contact with id=${id} not found`);
   }
@@ -27,14 +30,19 @@ export const deleteContact = async (req, res) => {
 };
 
 export const createContact = async (req, res) => {
-  const newContact = await contactsService.addContact(req.body);
+  const { id: owner } = req.user;
+  const newContact = await contactsService.addContact({ ...req.body, owner });
   res.status(201).json(newContact);
 };
 
 export const updateContact = async (req, res) => {
   const { id } = req.params;
-
-  const updatedContact = await contactsService.updateContact(id, req.body);
+  const { id: owner } = req.user;
+  const updatedContact = await contactsService.updateContact(
+    id,
+    owner,
+    req.body
+  );
   if (!updatedContact) {
     throw HttpError(404, `Contact with id=${id} not found`);
   }
@@ -44,9 +52,10 @@ export const updateContact = async (req, res) => {
 
 export const updateFavoriteContact = async (req, res) => {
   const { id } = req.params;
-
+  const { id: owner } = req.user;
   const updatedFavoriteContact = await contactsService.updateFavoriteContact(
     id,
+    owner,
     req.body
   );
   if (!updatedFavoriteContact) {
